pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openintent"
version = "0.4.0"
description = "Python SDK and Server for the OpenIntent Coordination Protocol"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "OpenIntent Contributors", email = "contributors@openintent.ai"}
]
keywords = [
    "openintent",
    "ai",
    "agents",
    "coordination",
    "protocol",
    "intent",
    "multi-agent",
    "llm",
    "server",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "sse-starlette>=1.6.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
# LLM Provider Adapters (install only what you need)
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
gemini = [
    "google-generativeai>=0.4.0",
]
grok = [
    "openai>=1.0.0",  # Grok uses OpenAI-compatible API
]
deepseek = [
    "openai>=1.0.0",  # DeepSeek uses OpenAI-compatible API
]
# Convenience extras
all-adapters = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
]
all = [
    "openintent[server,dev,all-adapters]",
]

[project.scripts]
openintent-server = "openintent.server.cli:main"
openintent = "openintent.cli:main"

[project.urls]
Homepage = "https://openintent.ai"
Documentation = "https://openintent.ai/docs"
Repository = "https://github.com/openintent-ai/openintent"
Changelog = "https://github.com/openintent-ai/openintent/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["openintent*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
